refactor:user_id값 바로 받을 수 있게끔 user_info에서 user_id만 추출하는 메소드 의존성주입
