+++ app/api/users.py
@@ -1,11 +1,11 @@
 from fastapi import APIRouter, Depends
 
-from app.core.auth import verify_token
+from app.core.auth import get_user_id, verify_token
 from app.core.base import BaseResponse
 from app.core.database import get_db
 from app.core.exception import ERROR_DUPLE, ERROR_UNKNOWN
 from app.schema.users import ModifyUserInfoRequestModel, UserOnboardRequestModel
-from app.services.users import UserService
+from app.services.user_service import UserService
 from app.utils.conveter import user_info_to_id
 
 router = APIRouter(prefix="/users", tags=["users"])
@@ -25,12 +25,11 @@ router = APIRouter(prefix="/users", tags=["users"])
 )
 async def complete_onboarding(
     req: UserOnboardRequestModel,
-    user_info=Depends(verify_token),
+    user_id=Depends(get_user_id),
     db=Depends(get_db),
 ):
     """유저의 온보딩완료 처리하는 API(취향설정, 닉네임 설정)"""
 
-    user_id = user_info_to_id(user_info)
     a, b, c, f, nickname = (
         req.atmospheres,
         req.bread_types,
@@ -68,10 +67,9 @@ async def complete_onboarding(
     """,
 )
 async def modify_user_info(
-    req: ModifyUserInfoRequestModel, user_info=Depends(verify_token), db=Depends(get_db)
+    req: ModifyUserInfoRequestModel, user_id=Depends(get_user_id), db=Depends(get_db)
 ):
     """유저 정보 수정하는 API"""
 
-    user_id = user_info_to_id(user_info)
     await UserService(db=db).modify_user_info(user_id=user_id, req=req)
     return BaseResponse(message="유저정보 수정 성공")
