+++ app/services/bakery_service.py
@@ -11,10 +11,8 @@ class BakeryService:
     def __init__(self, db: Session) -> None:
         self.db = db
 
-    async def get_bakeries_by_personal(self, area_code: str, user_info):
+    async def get_bakeries_by_personal(self, area_code: str, user_id: int):
         """유저의 취향이 반영된 빵집 조회하는 비즈니스 로직."""
-
-        user_id = user_info_to_id(user_info)
 
         # 구분자로 받은 지역코드 list로 반환
         area_codes = parse_area_codes(area_code)
